implementation of RBAC and added refresh token

+++ Inventory_management/app/crud.py
@@ -3,8 +3,10 @@ from sqlalchemy import func
 from datetime import date, timedelta
 from app.models import User, Category, Item
 from sqlalchemy.exc import IntegrityError
-from app.utils.auth import create_access_token
+from app.utils.auth import create_access_token,create_refresh_token
 from passlib.context import CryptContext
+from fastapi import HTTPException
+
 
 
 pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
@@ -26,7 +28,8 @@ def create_user(db: Session, user):
     db_user = User(
         name=user.name.strip(),
         email=user.email,
-        password=hashed_pw
+        password=hashed_pw,
+        role=user.role 
     )
 
     db.add(db_user)
@@ -78,11 +81,14 @@ def login_user(db, email: str, password: str):
     if not user:
         return None
 
-    token = create_access_token({"sub": str(user.id)})
+    token = create_access_token({"sub": str(user.id),"role": user.role})
+    refresh_token=create_refresh_token({"sub":str(user.id)})
 
     return {
         "access_token": token,
-        "token_type": "bearer"
+       
+        "refresh_token":refresh_token,
+        "role": user.role
     }
 
 
@@ -129,6 +135,7 @@ def update_user(db: Session, user_id: int, data: dict):
         ValueError: If email already exists
     """
     user = get_user(db, user_id)
+    
     if not user:
         return None
 
@@ -197,24 +204,19 @@ def patch_user(db: Session, user_id: int, data: dict):
 
 
 def delete_user(db: Session, user_id: int):
-    """
-    Delete a user by ID.
+    
+    user = db.query(User).filter(User.id == user_id).first()
 
-    Args:
-        db (Session): Database session
-        user_id (int): User ID
-
-    Returns:
-        dict | None: Success message or None
-    """
-    user = get_user(db, user_id)
     if not user:
-        return None
+        raise HTTPException(status_code=404, detail="User not found")
+
+    if user.items:
+        raise HTTPException(status_code=400, detail="User has items. Cannot delete.")
+
 
     db.delete(user)
     db.commit()
-    return {"message": "User deleted"}
-
+    return {"message": "User deleted successfully"}
 
 def create_category(db: Session, data: dict):
     """
